+++ integrationtest/vm/simulator/iam2/test_iam2_project_basic_ops_scheduler.py
@@ -34,6 +34,7 @@ test_stub = test_lib.lib_get_test_stub()
 case_flavor = dict(project_admin=                   dict(target_role='project_admin'),
                    project_operator=                dict(target_role='project_operator'),
                    project_member=                  dict(target_role='project_member'),
+                   system_admin=                    dict(target_role='system_admin'),
                    )
 
 def test():
@@ -41,10 +42,11 @@ def test():
 
     flavor = case_flavor[os.environ.get('CASE_FLAVOR')]
     # 1 create project
-    project_name = 'test_project'
-    project = iam2_ops.create_iam2_project(project_name)
-    project_uuid = project.uuid
-    project_linked_account_uuid = project.linkedAccountUuid
+    if flavor['target_role'] != 'system_admin':
+        project_name = 'test_project'
+        project = iam2_ops.create_iam2_project(project_name)
+        project_uuid = project.uuid
+        project_linked_account_uuid = project.linkedAccountUuid
 
     if flavor['target_role'] == 'project_admin':
         # 2 create virtual id
@@ -87,6 +89,13 @@ def test():
 	# 4 login in project
 	#project_inv=iam2_ops.get_iam2_projects_of_virtual_id(plain_user_session_uuid)
 	project_login_uuid = iam2_ops.login_iam2_project(project_name, plain_user_session_uuid).uuid
+    elif flavor['target_role'] == 'system_admin':
+        username = "systemAdmin"
+        password = 'b109f3bbbc244eb82441917ed06d618b9008dd09b3befd1b5e07394c706a8bb980b1d7785e5976ec049b46df5f1326af5a2ea6d103fd07c95385ffab0cacbc86'
+        vid_tst_obj = test_vid.ZstackTestVid()
+        virtual_id_uuid = vid_tst_obj.get_vid().uuid
+        test_stub.create_system_admin(username, password, vid_tst_obj)
+        project_login_uuid = acc_ops.login_by_account(username, password)
 
 
     # Image related ops: Add, Delete, Expunge, sync image size, Update QGA, delete, expunge
@@ -97,14 +106,17 @@ def test():
 
     vm_creation_option = test_util.VmOption()
     l3_net_uuid = test_lib.lib_get_l3_by_name(os.environ.get('l3VlanNetwork3')).uuid
-    acc_ops.share_resources([project_linked_account_uuid], [l3_net_uuid])
+    if flavor['target_role'] != 'system_admin':
+        acc_ops.share_resources([project_linked_account_uuid], [l3_net_uuid])
     vm_creation_option.set_l3_uuids([l3_net_uuid])
     image_uuid = test_lib.lib_get_image_by_name("centos").uuid
     vm_creation_option.set_image_uuid(image_uuid)
-    acc_ops.share_resources([project_linked_account_uuid], [image_uuid])
+    if flavor['target_role'] != 'system_admin':
+        acc_ops.share_resources([project_linked_account_uuid], [image_uuid])
     instance_offering_uuid = test_lib.lib_get_instance_offering_by_name(os.environ.get('instanceOfferingName_s')).uuid
     vm_creation_option.set_instance_offering_uuid(instance_offering_uuid)
-    acc_ops.share_resources([project_linked_account_uuid], [instance_offering_uuid])
+    if flavor['target_role'] != 'system_admin':
+        acc_ops.share_resources([project_linked_account_uuid], [instance_offering_uuid])
     vm_creation_option.set_name('vm_for_project_management')
     vm_creation_option.set_session_uuid(project_login_uuid)
     vm = test_stub.create_vm(image_uuid = image_uuid, session_uuid=project_login_uuid) 
@@ -134,8 +146,9 @@ def test():
     if plain_user_uuid != None:
         iam2_ops.delete_iam2_virtual_id(plain_user_uuid)
 
-    iam2_ops.delete_iam2_project(project_uuid)
-    iam2_ops.expunge_iam2_project(project_uuid)
+    if flavor['target_role'] != 'system_admin':
+        iam2_ops.delete_iam2_project(project_uuid)
+        iam2_ops.expunge_iam2_project(project_uuid)
 
     test_util.test_pass('success test iam2 login in by project admin!')
